pyproject.toml
[project]
name = "agents-memory"
version = "0.2.0"
description = "Agent memory patterns with Mem0, Groq, and open-source tools"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "strands-agents>=1.0.0",
    "strands-agents-tools>=0.1.0",
    "litellm>=1.0.0",
    "mem0ai>=0.1.0",
    "sentence-transformers>=2.0.0",
    "qdrant-client>=1.0.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "numpy>=1.26.0",
    "scikit-learn>=1.4.0",
    "rich>=13.0.0",
    "duckduckgo-search>=5.0.0",
    "fastapi>=0.135.1",
    "uvicorn[standard]>=0.30.0",
    "braintrust[otel]>=0.0.154",
    "anthropic>=0.39.0",
    "ddgs>=1.0.0",
    "opentelemetry-api>=1.21.0",
    "opentelemetry-sdk>=1.21.0",
]

[build-system]
requires = [
    "hatchling",
]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/agents_memory"]
